+++ tests/entrypoints/pooling/scoring/test_cross_encoder_correctness_mteb.py
