+++ tests/entrypoints/pooling/scoring/test_cross_encoder_online_vision.py
